refacto: create record_api
