+++ Holovibes/includes/API.hxx
@@ -261,36 +261,6 @@ inline void set_import_type(ImportType value) { UPDATE_SETTING(ImportType, value
 
 inline CameraKind get_camera_kind() { return GET_SETTING(CameraKind); }
 inline void set_camera_kind(CameraKind value) { UPDATE_SETTING(CameraKind, value); }
-/*! \} */
-
-/*!
- * \name Recording
- * \{
- */
-inline holovibes::Device get_record_queue_location() { return GET_SETTING(RecordQueueLocation); }
-
-inline uint get_record_buffer_size() { return static_cast<uint>(GET_SETTING(RecordBufferSize)); }
-
-inline std::string get_record_file_path() { return GET_SETTING(RecordFilePath); }
-inline void set_record_file_path(std::string value) { UPDATE_SETTING(RecordFilePath, value); }
-
-inline std::optional<size_t> get_record_frame_count() { return GET_SETTING(RecordFrameCount); }
-inline void set_record_frame_count(std::optional<size_t> value) { UPDATE_SETTING(RecordFrameCount, value); }
-
-inline RecordMode get_record_mode() { return GET_SETTING(RecordMode); }
-inline void set_record_mode(RecordMode value) { UPDATE_SETTING(RecordMode, value); }
-
-inline size_t get_record_frame_skip() { return GET_SETTING(RecordFrameSkip); }
-inline void set_record_frame_skip(size_t value) { UPDATE_SETTING(RecordFrameSkip, value); }
-
-inline bool get_frame_record_enabled() { return GET_SETTING(FrameRecordEnabled); }
-inline void set_frame_record_enabled(bool value) { UPDATE_SETTING(FrameRecordEnabled, value); }
-
-inline bool get_chart_record_enabled() { return GET_SETTING(ChartRecordEnabled); }
-inline void set_chart_record_enabled(bool value) { UPDATE_SETTING(ChartRecordEnabled, value); }
-
-inline uint get_nb_frame_skip() { return GET_SETTING(FrameSkip); }
-inline uint get_mp4_fps() { return GET_SETTING(Mp4Fps); }
 
 inline uint get_camera_fps() { return GET_SETTING(CameraFps); }
 inline void set_camera_fps(uint value) { UPDATE_SETTING(CameraFps, value); }
